Multilingual_VirtualAI_Google_gTTs/requirements.txt
SpeechRecognition
pipwin
pyaudio
gTTs
google-generativeai
python-dotenv
streamlit

-e .